mindscape/bioinformatics-workflow-engine/requirements.txt
# requirements.txt

ruamel.yaml
pytest
pandas
numpy
scipy
matplotlib
snakemake
```